Replace RunInterval/SleepInterval with IntervalAGAIN/IntervalDONE
RunInterval goes away to keep down feature creep, we can always add it back later if we ever need it.
SleepInterval is now named IntervalAGAIN, which is the time we sleep between each cron job execution, i.e. between each db txn commit to spread the load.
The new IntervalDONE is how long time to sleep after the cron job has completed and returned 'DONE' to indicate there is no more work.
If IntervalDONE is NULL this means we won't execute the cron job every again if it returns 'DONE', which is also the default.

The IntervalAGAIN and IntervalDONE are useful for cron jobs that most of the times have very little to do,
but at times have a lot to do and needs to be called with very short intervals in between each db txn,
e.g. maybe a 00:00:00.05 IntervalAGAIN and a 00:00:01 IntervalDONE, which would create a cron job
that sleeps 1 second between each run, but when it has job to do and returns AGAIN is called again after 50 ms.