Add support for connection pooling
Example:
SELECT cron.New_Connection_Pool(_Name := 'My pool', _MaxProcesses := 2);
SELECT cron.Register('Foo(integer)', _ConnectionPool := 'My pool');
SELECT cron.Register('Bar(integer)', _ConnectionPool := 'My pool');
SELECT cron.Register('Baz(integer)', _ConnectionPool := 'My pool');

This would limit the three cron job processes to only run two concurrently.

If the MaxProcesses is reached, the cron job processes will not run until
one of the processes within the same connection pool has finished.

The cron job process in a connection pool that has been waiting the longest
has priority and will run as soon as it is allowed to run, i.e. when some
other already running process in the connection pool finishes.

If _ConnectionPool is not specified (or set to NULL) then the cron job
is guaranteed to run as many processes as specified using _Processes,
which by default is 1.

The purpose of the cron.ConnectionPools columns LastCycleAt,ThisCycleAt
and CycleFirstProcessID is to make it easier to measure how long time
it takes until all cron job processes in a connection pool have been executed.

This is useful when having a "Non-Time Critical But Important" type of
connection pool where it's not a problem if a job is delayed a few minutes or hours,
but where it is a problem if it wouldn't run for an entire day.

This can help a sysadmin to regulate the value of MaxProcesses.
If it's e.g. is 1 to begin with, and the sysadmin notice the cycle
is ~24 hours, then by increasing the MaxProcesses from 1 to 2,
the cycle should go down to ~12 hours.

The goal is of course of keep MaxProcesses as low as possible for all
non-time critical jobs, to not waste any database resources for no good reason.

This commit also makes start-up a bit safer by making sure no other
cron job processes are running before starting up again, to prevent
problems if pgcronjob would be killed or crash.