Log when other db queries are waiting to the new table cron.LogWaiting
And if we've never seen the (NumPIDs,MaxSeconds) combination before,
log the entire output from pg_stat_activity to the new table LogWaitingPgStatActivity.

Since we determine if things are waiting in one query, and log what's waiting in a separate db txn,
there is a risk we won't catch what was waiting, but this is not a problem since we only care about
things are have been waiting for multiple seconds anyway.