Clarify how $ConnectTime is measured by using a new intermediate variable, $StartupTime
The $StartupTime is the time() when the program was started
just before connecting. It's set once and won't change,
so it can be reused at other places if necessary.

The $ConnectTime is the time it took to DBI->connect,
measured as the different between time() and the $StartupTime.

The $ConnectTime is useful to know whether or not it will
be better to close the connection to reduce the number
of concurrently open pg connections, or if it's better
to keep it open in case we will need to run the cron process
again within a shorter time than it would take to re-connect.