Change all if/elsif-statements to explicitly check the value
While it can be convenient to just check "if ($Foo)",
it's arguably safer and more clear to check "if ($Foo == 1)"
or whatever value you expect $Foo to have.

The $$RunAtTime variable is set to 0 to not enter the if block
that before just checked "if ($$RunAtTime)". Now this stands out
since it now checks "if ($$RunAtTime > 0)". Otherwise someone
might think "if we set $$RunAtTime to zero it will for sure run
since that's way before time()", but it wouldn't since 0 means
"process is already running".

I didn't change "elsif ($$Run->pg_ready)" since the
the official documentation of DBD::Pg doesn't say anything
about the returned value, expect that it will eval to true.