Recover from some common expected database error during cron.Run() instead of dying
We will still die for all database errors on the separate database connections
reserved for LISTEN, cron.Dispatch(), etc. This only affects the database connections
reserved for cron.Run(), e.g. the connections in which we execute the cron jobs.

If we would get an error in the general database connections, we will still die()
thanks to RaiseError => 1 in those two connections.

The new function CheckErr() takes the return value from the DBI call as first arg,
and the $ProcessID as second arg.

Since you cannot tell just by the return value if a DBI call was an error or not,
since you have to check if it's undef AND $dbh->err, this handling has been
centralized and put into CheckErr().

CheckErr() returns 0 if it's OK, just like DBI's connect(), so you can use it
in the same way, e.g. "CheckErr() or die;" or maybe more useful "CheckErr() or next;",
to just skip to the next in the loop.

A new table ErrorLog has been added to which all database error will be written,
handled or not.

For any uncommon error, we will still die() allowing pgcronjob to be restarted.

If some unhandled error is common enough, we might add more of them to the list of handled errors.