Always proceed regardless of whatever value cron.Processes.RunAtTime has
The pgcronjob Perl program might have a different system time than the database,
so let the Perl program decide on its own when it's time to run a cron job.
Before, cron.Run() would simply return if RunAtTime was in the future or NULL.

This also means cron.Schedule() won't have to update cron.Processes.RunAtTime,
but can just send the pg notification to tell pgcronjob to run the job ASAP.

This is also good for concurrency since the cron.Processes row might be locked by
pgcronjob, but that potential problem goes away now since we don't need to update
the row.