Don't include the $ProcessID in the prepared SQL query, use $1 placeholder instead.
This allows the same prepared query to be shared by all processes.
This is a preparatory change to implement sharing of connections between
processes in the same connection pool.

Normal processes that don't have a connection pool will still have a dedicated
connection, that is kept open if it should run very soon again, or closed if not.